[UPD] feat(email-attestation): Attestation for promoters with links to bitstreams.
Refactored `AttestationEmailAction` class in order to send emails to both authors and promoters.
Promoters and authors receive both different mails. In the promoter's email, there are access urls to download bitstreams.

Added a new controller which handles a new endpoint to download bitstream. This endpoint can be used by the promoter to retrieve a specific bitstream.
A hash needs to be given in the request params. This hash must correspond to the hash of one of the promoter's email in order to download the bitstream's file.

Authored-By: Michaël Pourbaix <michael.pourbaix@uclouvain.be>