[ADD] feat(workflow): Additional action to accept an item without diffusion of its bitstreams.
Modified the workflow action class to handle a new option: 'Accept without diffusion'.
This allows the manager to see a new button in the frontend which will trigger the behaviour.

When this option is selected by the manager, it will restrict the acces of all bitstream to admin only.
Also the manager can leave a message which will be saved in the 'dc.description.diffusion' metadata field of the item.

This commit is linked to another one in the frontend which creates the corresponding button.

Authored-By: Michaël Pourbaix <michael.pourbaix@uclouvain.be>