[ADD] feat(file-loader) File loader class for configuration files
Added a new file loader class that allows to load a file and read it.
It can also be used to determine if the file has change since the last time it was red.
By using this class to load files content, it allow us to modify config files without restarting the application.

Authored-By: Michaël Pourbaix <michael.pourbaix@uclouvain.be>