[UPD] feat(backend-fixes): Global search, profile edit and profile file upload.
[Master Thesis specific]
-- Global Search --
Adding a new condition into the Solr query for the global search: 'entityType:MasterThesis'.
This forces Solr to only return Item with a 'entityType' of 'MasterThesis' and so it does not return Person items.

-- Profile Edit --
A user can no longer edit his profile. This prevents the manager from changing his managed degree codes.

-- Profile File Upload --
We no longer need to upload a file when creating/modifying a user researcher profile.
Done by modifying the form for this functionnality called 'person-edit'.

This commit is specific for the master thesis repository and is then separated from the others.

Authored-By: Michaël Pourbaix <michael.pourbaix@uclouvain.be>