[ADD] feat(general): File loader class for configuration files
Added a new file loader class that allows to load a file and read it.
Each time the file is read there is a version check just before loading the data.
By using this class to load files content, it allows us to modify config files without restarting the application.

Authored-By: Michaël Pourbaix <michael.pourbaix@uclouvain.be>