[ADD] feat(submission): Change request form section
When a manager requests a 'return to submitter', his reason is store in the item's metadata.
A new submission section displays the current change request for the submitter by reading the corresponding metadata.
Also, a new workflow step removes the current active change request each time an item is submitted.

@Authored-By: Michaël Pourbaix <michael.pourbaix@uclouvain.be>