[UPD] feat(affiliationTree) Optimization to get rid of backend cache.
Since backend cache was complex and made our life harder, this commit is optimizing the operations to decrease the base response time and get rid of this cache.
Instead of using the 'ItemService' we directly extract metadata from the retrieved Solr document.
This allows to go from a response time of '2500ms' avg. to '100ms' avg. (sometimes even less).

Changes:
-> Removed the caching system.
-> Using SolrDocument to extract metadata instead of using the database.
-> Optimizing some code + more error handling.
-> Deleted consumer that was generating an affiliation tree when an OrgUnit was modified.

Authored-by: Michaël Pourbaix <michael.pourbaix@uclouvain.be>