[ADD] feat(consumers): Consumer to clear 'filler' values like NA or N/A
A configurable consumer has been added that will check for values like NA or N/A in the item's metadata.
It is only triggered when the item is not in workspace.
In general, it will remove the 'filler' metadata values but if the field is in a form group, it will replace them with a placeholder value.

Authored-by: Michaël Pourbaix <michael.pourbaix@uclouvain.be>