[ADD] feat(authors-notification): Mail to notify authors of submission
When a publication is submitted, a mail is sent to all the authors.
I added a system to get the metadata of the publication in the mail depending on the given language.
This system keeps the metadata in the same order every time.
It can also translate a value using a given vocabulary name.

Authored-by: Michaël Pourbaix <michael.pourbaix@uclouvain.be>