[UPD] feat(init-script): 'authenticated-users' group can submit publications.
Nobody could submit in the publication collection appart from admins.
Added a new group to fix that. This group is configured to be able to submit in the 'Publication' collection.
The user 'user1@dspace.org' is a member of this group by default.
Also added an admin user to use for imports (import@dspace.org).

Authored-by: Michaël Pourbaix <michael.pourbaix@uclouvain.be>