[ADD] feat(submission): Allow restriction on submission
Allows to restrict a single submission for some collection.
If the current-logged user is member of the "COLLECTION_{uuid}_UNRESTRICTED_SUBMISSION", it can create multiple submission, otherwise a single submission is allowed.
If this group doesn't exists, no restrictions could be found; the default DSpace behavior is used (multiple submission allowed)

Authored-by: Renaud Michotte <renaud.michotte@uclouvain.be>