[FIX] feat(setup) Fixed consumer error at setup.
The consumer was triggered at the setup phase when the 'Journal' collection does not already exist and it throws an error.
To fix that I have added a check for the list of IDs to process in 'JournalDefaultDeletePolicyConsumer'.
This prevents from entering the 'end()' method when no journals need to be processed.
So 'No journals => No collection' is OK but 'Some Jourals => No collection' throws an error.

Authored-by: Michaël Pourbaix <michael.pourbaix@uclouvain.be>