[ADD] feat(item-validator): Validate items before creation/update.
Added a new system that allows to define a validator for each item entity-type.
When an item is created or updated, the validator is executed an can cancel the update if some requirements are not met.
For profiles, a new validator has been added to prevent duplicate identifiers (on email, fgs and ORCID).
Also added a new 'ExceptionHandler' to DSpace that allows to send a specfic translatable string to the frontend.

Authored-by: Michaël Pourbaix <michael.pourbaix@uclouvain.be>